[ADD] Form Request for storing GuessByImage type card. [FIX] Display cards of a collection.
Rule class for StoreGuessByImage type card has been created as well as its own controller class with an update method inside.
Sent previous page links to the form as to allow user to step back from the form.
Changed resource (toArray()) for card so that the data can be accessed more straightforward through an API call.
Created an image collection as well as an image resource so that a more detailed data can be retrieved from an API call (e.g. URL, type, path, created_at...) useful when loading files with JavaScript File API.
Allowed the extraction of the image collection when retrieving a card resource.
Removed useless code or test code.
The method that removes card assets is now triggered whenever a card is deleted through the `deleting` observer.
Included some styles for the card update form.
Ignored some folder insides the public folder.
Added some translations for the card update form.
The method used by the card update form is now a spoofed `PUT` instead of a `POST`.