[UPDATE] Collection card list UI [FIX] Card storing fails when no images is attached.
The default number of cards displayed by default in the view has been changed to a more suitable one to its purpose.

The "store()" method checks if an image has been attached.

The grid of the card list has been set so that the card items keeps the same dimensions even with screen resizing.

Applied the "Show more" button functionality for the displayed cards. This way once clicked, it will load the next coming cards while keeping the one's that had already been displayed.

The "Show more" button only displays when there are more cards to load, if not it won't show.

Removed useless or testing code.

The "getCards()" method has been modified in order to push the data retrieved from the ajax response instead of setting them in the state. Thus, whenever a new page of cards is loaded it won't overwrite the state but just will add new cards at the end of the array.

Removed the "container" class for the "main" component so to improve grid customization of the children components.